:recycle: API通信に関する処理をviewsに移動
