+++ frontend/src/components/templates/AddScheduleDialog.tsx
@@ -4,7 +4,11 @@ import { FormDialog } from "../parts/FormDialog";
 import { MonthContext } from "@/provider/CalendarProvider";
 import { client } from "@/libs/api/axios";
 
-export const AddScheduleDialog = () => {
+type Props = {
+  handleSaveSchedule: any;
+};
+
+export const AddScheduleDialog = ({ handleSaveSchedule }: Props) => {
   const {
     schedule,
     daySelected,
@@ -12,33 +16,9 @@ export const AddScheduleDialog = () => {
     setSchedule,
     setDaySelected,
     setShowAddDialog,
-    setSchedules,
   } = useContext(MonthContext);
 
   const handleClose = () => {
-    setSchedule({
-      id: "",
-      title: "",
-      date: daySelected,
-      description: "",
-      location: "",
-    });
-    setShowAddDialog(false);
-  };
-
-  // TODO viewsに移動
-  const handleSaveSchedule = async () => {
-    const body = {
-      title: schedule.title,
-      date: daySelected.toISOString(),
-      location: schedule.location,
-      description: schedule.description,
-    };
-    await client.post("schedule/", body);
-    client.get("schedule/").then(({ data }) => {
-      setSchedules(data);
-    });
-
     setSchedule({
       id: "",
       title: "",
